fix config pkg dependency Signed-off-by: Marco <llan@redhat.com>
